α0.0.4 patch 4 multi-commit 1
* Achievement line swap fix

* Achievement 51 condition fix

* Improved word generation on large delay

If thought less than 100,000 times, purely iterate every time at a probability. Else, calculate the binary probability distribution for up to 2.500 million times, multiply by a coefficient so that the number of thoughts become the actual number. This is to prevent the game from freezing for an extended amount of time when if the game needs to calculate say for an year, equivalent to 31.558 million seconds.

* Update changelog.html